PlatinumTier/requirements.txt
groq>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0
watchdog>=3.0.0
playwright>=1.44.0
google-api-python-client>=2.131.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.0
httpx>=0.27.0
tweepy>=4.14.0
